Add SharedFixture for integration tests to share DB across tests classes
The changes introduce a new collection fixture named "SharedFixture" for integration tests in the `AuctionService.IntegrationTests` namespace. This fixture is defined in the `SharedFixture.cs` file and is used to share the `CustomWebAppFactory` instance across multiple test classes.

In `AuctionBusTests.cs` and `AuctionControllerTests.cs`, the `IClassFixture<CustomWebAppFactory>` interface is removed, and the `[Collection("SharedFixture")]` attribute is added to the test classes. This change ensures that the `CustomWebAppFactory` instance is shared among tests in these classes, promoting better resource management and potentially reducing test execution time.