Add Kubernetes configs for bid-svc Deployment and Service
Updated bid-depl.yml to include Deployment and Service for the
bid-svc application. The Deployment specifies the container
image viniciussmv/bid-svc, sets the image pull policy to Never,
and uses environment variables from the bid-svc-config ConfigMap.
The Service is defined as ClusterIP to expose the application
internally on port 80.

Updated config.yml to define a new ConfigMap named bid-svc-config.
This ConfigMap includes environment variables such as ASP.NET Core
URLs, RabbitMQ credentials, MongoDB connection string, and URLs
for the Identity and Auction services, necessary for the bid-svc
application to function correctly.